demos/requirements.txt
maestro-sdk
fastapi>=0.104.0
uvicorn[standard]>=0.24.0